Initial port to Linux, and
I FINALLY FOUND THE WEIRD MEMORY BUG!
When I introduced the IPC mechanism, I had to change where the datafile was
closed so the parent could access it after the child exits. However, I
forgot that I had specified the file buffer directly, and this was
still being freed. Unfortunately, efence was only marginally useful in finding
this because memory accesses inside the file io functions are somehow
trapped differently. But, 4 hours later.... I finally figured it out.

jeff


git-svn-id: https://svn.internet2.edu/svn/owamp/trunk@808 a2d82226-602c-0410-bba7-e51b83dd10d2

+++ owamp/capi.c
@@ -1605,7 +1605,7 @@ OWPFetchSession(
             *err_ret = OWPErrFATAL;
             goto failure;
         }
-        if(dowrite && ( fwrite(buf,1,toff,fp) != toff)){
+        if(dowrite && ( fwrite(buf,1,toff,fp) != (size_t)toff)){
             OWPError(cntrl->ctx,OWPErrFATAL,errno,
                     "OWPFetchSession: fwrite(): %M");
             dowrite = False;
